movieList HOC
